[6868] Extract AbstractValidatingAggregateRoot to a new sirius-web-core-domain
Bug: https://github.com/eclipse-sirius/sirius-web/issues/6868
