+++ packages/core/backend/pom.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
- Copyright (c) 2022, 2025 Obeo.
+ Copyright (c) 2022, 2026 Obeo.
  This program and the accompanying materials
  are made available under the terms of the Eclipse Public License v2.0
  which accompanies this distribution, and is available at
@@ -40,5 +40,6 @@
 		<module>sirius-components-core-graphql</module>
 		<module>sirius-components-collaborative</module>
 		<module>sirius-components-graphql-api</module>
+		<module>sirius-web-core-domain</module>
 	</modules>
 </project>
