+++ packages/core/backend/sirius-web-core-domain/.settings/org.springframework.ide.eclipse.boot.prefs
@@ -1,0 +1,2 @@
+eclipse.preferences.version=1
+spring.boot.launch.profile.history=dev;
