+++ packages/core/backend/sirius-web-core-domain/.settings/org.springframework.ide.eclipse.prefs
@@ -1,0 +1,2 @@
+boot.validation.initialized=true
+eclipse.preferences.version=1
