+++ packages/core/backend/sirius-web-core-domain/src/test/resources/logback-test.xml
@@ -1,0 +1,3 @@
+<configuration>
+  <logger name="com.tngtech.archunit" level="INFO" />
+</configuration>
