+++ packages/sirius-web/backend/sirius-web-domain/src/main/java/org/eclipse/sirius/web/domain/boundedcontexts/image/Image.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2024 Obeo.
+ * Copyright (c) 2024, 2026 Obeo.
  * This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v2.0
  * which accompanies this distribution, and is available at
@@ -17,7 +17,7 @@ import java.util.Objects;
 import java.util.UUID;
 
 import org.eclipse.sirius.components.events.ICause;
-import org.eclipse.sirius.web.domain.boundedcontexts.AbstractValidatingAggregateRoot;
+import org.eclipse.sirius.web.core.domain.AbstractValidatingAggregateRoot;
 import org.eclipse.sirius.web.domain.boundedcontexts.image.events.ImageCreatedEvent;
 import org.eclipse.sirius.web.domain.boundedcontexts.image.events.ImageDeletedEvent;
 import org.springframework.data.annotation.Id;
